Added Items and Crafting xml
Item list and crafting recipe is now store in an xml file. Program looks
at this xml for the type of item and how the item can be crafted. This
is a constant xml and will not be changed by game play or the user (aka
is not a place to save the users inventory). Basically a read-only file.