Inventory Complete :)
The inventory is now finished. It finds the items that the user has, and
compares them to the itemID in items.xml so it can find the info about
the item. The items can also stack (if the item is stackable). All that
is left is to integrate this into the UI. Right now, it just prints out
the inventory into the console.