AI even better
Follows you without getting stuck (99% of the time it wont get stuck).
The AI now follows the players history path. Still one more bug left,
which is when the player has no history (for example, when the game just
starts and the player hasn't move yet), the ai has no history to follow
so it gets confused